% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\docType{package}
\name{dependenciesMap}
\alias{dependenciesMap}
\title{dependenciesMap: A package that allows to create objects in order to map dependencies in scripts}
\description{
Each object of this package is used to map an input or an output that can be :\itemize{
\item A file (RData, csv, xlsx) : \code{\link{Input_file}} and \code{\link{Output_file}}
\item A db table : \code{\link{Input_db}} and \code{\link{Output_db}}
\item A dataframe : \code{\link{Input_df}} and \code{\link{Output_df}}
}
}
