% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_graph.R
\name{get_dependencies_graph}
\alias{get_dependencies_graph}
\title{Returns graph of dependencies for the main processes of relationships}
\usage{
get_dependencies_graph(relations, processes, view_time = T)
}
\arguments{
\item{relations}{data frame with dependencies (with fields from, to and time)}

\item{processes}{data frame with main processes involved (with parameters from, to and time)}

\item{view_time}{Boolean : True (by default), if time as labels of edges shown, False otherwise}
}
\value{
Returns graph of dependencies for the main processes of relationships
}
\description{
Returns graph of dependencies for the main processes of relationships
}
