% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\name{getProcesses}
\alias{getProcesses}
\title{Gets the names of the processes used in the input and output objects}
\usage{
getProcesses()
}
\description{
Gets the names of the processes used in the input and output objects
}
