% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\name{init_dependencies}
\alias{init_dependencies}
\title{Initializes objects that will record inputs, outputs and dependencies}
\usage{
init_dependencies()
}
\description{
Initializes objects that will record inputs, outputs and dependencies
}
