% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\docType{class}
\name{Input_db-class}
\alias{Input_db-class}
\alias{Input_db}
\title{Inputs loaded in the data preparation program from data base's tables}
\description{
Inputs loaded in the data preparation program from data base's tables
}
\section{Fields}{

\describe{
\item{\code{conn}}{: connection of the data base}

\item{\code{query}}{: query}

\item{\code{process}}{: process where input is loaded in the data preparation program (to be added by user)}

\item{\code{table}}{: name of the data base's table used to get the data (to be added by the user)}
}}

\section{Methods}{

\describe{
\item{\code{f_dbGetQuery()}}{function \code{dbGetQuery()} from DBI package

                             By default : \code{dbGetQuery( conn, query )}

                             \subsection{Example} { \code{par_loc <- dbGetQuery( SCDatabase, 'select * from PAR_LOC_geoMapping' ) } is now replaced by

                            \code{par_loc <- Input_db( conn = SCDataBase, query = 'select * from PAR_LOC_geoMapping', process = 'SLA', table = 'PAR_LOC_geoMapping')$f_dbGetQuery()  } }}
}}

