% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\docType{class}
\name{Input_df-class}
\alias{Input_df-class}
\alias{Input_df}
\title{Inputs loaded in the data preparation program from data frames previously generated}
\description{
Inputs loaded in the data preparation program from data frames previously generated
}
\section{Fields}{

\describe{
\item{\code{name}}{: name of the data frame loaded (to be added by user)}

\item{\code{data}}{: data loaded (optional attribute)}

\item{\code{process}}{: process where input is loaded in the data preparation program (to be added by user)}
}}

\section{Methods}{

\describe{
\item{\code{f_add()}}{does nothing, just allows to keep track of the new data frame used as input

                                    to be used when in a process, a new data frame is used and there is no need to return it

                                    \subsection{Example}{ \code{Input_df( name = 'MasterData', data = MasterData, process = 'Future_Forecast' )$f_add()}}}

\item{\code{f_return()}}{simply returns the data of the object 

                                   By default : \code{return(data)}

                                   \subsection{Example}{\code{MasterData2 <- MasterData}  is now replaced by

                                   \code{MasterData2 <- Input_df( name = 'MasterData', data = MasterData, process = 'Future_Forecast' )$f_return()} }}
}}

