% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\docType{class}
\name{Input_file-class}
\alias{Input_file-class}
\alias{Input_file}
\title{Inputs loaded in the data preparation program from files}
\description{
Inputs loaded in the data preparation program from files
}
\section{Fields}{

\describe{
\item{\code{file}}{: name of the file loaded (R.data, csv, xlsx)}

\item{\code{path}}{: path of file (NULL if omitted)}

\item{\code{version}}{: version in path ( DEV/ for Development, blank for production) (NULL if omitted)}

\item{\code{process}}{: process where input is loaded in the data preparation program (to be added by user)}
}}

\section{Methods}{

\describe{
\item{\code{f_load(...)}}{function \code{load()} 

                               By defaut : \code{load( paste0( path, version, file ), envir = .GlobalEnv, ...)} 

                               \subsection{Example}{\code{load( 'SalesnFcst.Rdata' )} is now replaced by

                               \code{Input_file( file = 'SalesnFcst.Rdata', process = 'SalesnFcst' )$f_load()}  (version and path = NULL)}}

\item{\code{f_read.csv(...)}}{function \code{read.csv()} 

                              By default : \code{read.csv( paste0( path, version, file ), ... )} 

                              \subsection{Example}{\code{read.csv( 'Dependencies_Map/dependencies.csv' )} is now replaced by

                              \code{Input_file( file = 'dependencies.csv', path = 'Dependencies_Map/', process = 'SalesnFcst')$f_read.csv()}  (version = NULL) }}

\item{\code{f_read.xls(...)}}{function \code{read.xls()} from gdata package

                               By default : \code{gdata::read.xls( paste0( path, version, file ), ... )}}

\item{\code{f_read.xlsx(...)}}{function \code{read.xlsx()} from openxlsx package

                               By default : \code{openxlsx::read.xlsx( paste0( path, version, file ), ... )} 

                               \subsection{Example}{\code{read.xlsx( paste0( PathLoad, Version, 'Baselines_and_forecastable.xlsx' ), sheet = 'SkuMatrix', colNames = TRUE  )} is now replaced by

                               \code{Input_file( file = 'Baselines_and_forecastable.xlsx', path = PathLoad, version = Version, process = 'SalesnFcst' )$f_read.xlsx( sheet = 'SkuMatrix', colNames = TRUE )} }}
}}

