% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\docType{class}
\name{Output_df-class}
\alias{Output_df-class}
\alias{Output_df}
\title{Outputs in the data preparation program from data frames previously generated (can be used to modified a global variable)}
\description{
Outputs in the data preparation program from data frames previously generated (can be used to modified a global variable)
}
\section{Fields}{

\describe{
\item{\code{name}}{: name of the data frame (to be added by user)}

\item{\code{data}}{: data  generated (NULL if omitted)}

\item{\code{process}}{: process where output is generated in the data preparation program (to be added by user)}
}}

\section{Methods}{

\describe{
\item{\code{f_add()}}{does nothing, just allows to keep track of the new data frame created as output

                                    to be used when in a process, a new data frame is created and there is no need to return it

                                    \subsection{Example}{ \code{Output_df( name = 'MasterData', data = new_data, process = 'CoV' )$f_add()} }
                                    }

\item{\code{f_return(...)}}{simply returns the data of the object :

                                     By default: \code{return(data)}

                                     \subsection{Example} { \code{MasterData <- left_join( MasterData, Adj_CoV_Result, by = c('DMDUNIT','LOC','DMDGROUP','STARTDATE') ) } is now replace by

                                    ( \code{new_data <- left_join( MasterData, Adj_CoV_Result, by = c('DMDUNIT','LOC','DMDGROUP','STARTDATE'))} )

                                    \code{MasterData <- Output_df( name = 'MasterData', data = new_data, process = 'CoV' )$f_return()} }}
}}

