% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies_code.R
\docType{class}
\name{Output_file-class}
\alias{Output_file-class}
\alias{Output_file}
\title{Output files saved in the data preparation program}
\description{
Output files saved in the data preparation program
}
\section{Fields}{

\describe{
\item{\code{file}}{: name of the file that is going to be saved (R.data, csv, xlsx)}

\item{\code{data}}{: data that is going to be saved}

\item{\code{data_name}}{: name of the variable that contains the data (string)}

\item{\code{path}}{: path of file (NULL if omitted)}

\item{\code{version}}{: version in path ('DEV/' for Development, '' for production) (NULL if omitted)}

\item{\code{process}}{: process where input is loaded in the Data preparation program (to be added by user)}
}}

\section{Methods}{

\describe{
\item{\code{f_save(...)}}{function \code{save()} 

                              By default : \code{save( data, paste0(path, version, file), ... )}

                               \subsection{Example}{\code{save( MasterData, file =  paste0( PathLoad, Version, 'SalesnFcst.Rdata' ) )} is now replaced by

                               \code{Output_file(data = MasterData, data_name = 'MasterData', file = 'SalesnFcst.Rdata', path = PathLoad, version = Version, process = 'SalesnFcst')$f_save()} }}

\item{\code{f_write.csv(...)}}{function \code{write.csv()} 

                              By default : \code{write.csv(data, paste0( path, version, file ), ... )} 

                              \subsection{Example}{\code{write.csv(FutFcst,paste0(PathLoad,'Future_Fcst.csv'), row.names = FALSE, na = '')} is now replaced by

                              \code{Output_file(data = FutFcst, file = 'Future_Fcst.csv', path = PathLoad, process = 'Forecast')$f_write.csv(row.names = FALSE, na = '')} }}

\item{\code{f_write.xlsx(...)}}{function \code{write.xlsx()} from openxlsx package

                               By default : \code{write.xlsx(data, paste0( path, version, file ), ... )} 

                              \subsection{Example}{\code{write.xlsx(FutFcst, paste0(PathLoad,'Future_Fcst.xlsx') )} is now replaced by

                              \code{Output_file(data = FutFcst, file = 'Future_Fcst.xlsx', path = PathLoad, process = 'Forecast')$f_write.xlsx()} }}
}}

